Add solution to blank screen problem to DOCS.md
I use Debian Trixie AMD64 with XFCE 4.18. When I updated to version 1.83.1 and opened the program, all I saw was a mangled version of the interface without text.
I tested various versions of VS Codium and VS Code to find when the issue originated, and found it started with Codium version 1.82 (didn't occur in 1.81.x, and couldn't replicate the bug in regular VS Code).
I was about to file a bug report in this project when I found that the same issue was occurring upstream due to a Chromium, and thus Electron update (see microsoft/vscode#190437). I found that it was an issue with Mesa GPU caching upon VS Code updates (as well as updates of some other Electron-based applications). The issue was solved by deleting the GPU cache (which would make sense why I couldn't replicate the issue in regular Code: since it was a new install, the GPU cache was also brand new).
In case other people come across the same issue while using VSCodium, I would like this solution to be available in the documentation until the Electron bug is quashed.