fix: update Electron to v34.3.3
This fixes #2250, in which VSCodium couldn't be installed on older
systems because of the used glibc version.

```
Error:
 Problem: conflicting requests
  - nothing provides libm.so.6(GLIBC_2.29)(64bit) needed by
    codium-1.99.32704-el9.x86_64 from @commandline
(try to add '--skip-broken' to skip uninstallable packages or '--nobest'
to use not only best candidate packages
```

The error comes from the currently used Electron version (v34.3.2).
https://github.com/electron/electron/issues/45884

The problem was fixed in the Electron Project and later backported.
Version v34.3.3 already includes the fix.
https://github.com/electron/electron/commit/ac5b5995e85e6f0b9bf5e82ff6df20d52c7d56bc